requirements.txt
pandas
openpyxl
duckdb
fastapi
uvicorn
chromadb
sentence-transformers
rank-bm25
pypdf
python-docx
beautifulsoup4
pyngrok
requests